requirements/base.txt
Django==2.0.2
django-model-utils==3.1.1
djangorestframework==3.7.7
